ILLDAN-140 [fix] : 타입오류를 해결한다

+++ src/main/java/server/poptato/todo/application/response/RoutineCountDto.java
@@ -1,4 +1,4 @@
 package server.poptato.todo.application.response;
 
-public record RoutineCountDto(String day, int count) {
+public record RoutineCountDto(String day, Long count) {
 }
